Update URL for Library Manager submission instructions
Arduino has created a dedicated repository for Library Manager submissions and moved the documentation there. The content of the page pointed to by the previous URL was replaced by a link pointing to the new Library Manager FAQ, so the user will still find the information after a few clicks, but it will be a better experience for them to be able to get to the instructions with a single click.